separate files and improve readme
