+++ src/challenges/js/30DaysJs-30/src/User.js
@@ -1,0 +1,18 @@
+export class User {
+  constructor(name) {
+    this.name = name;
+    this.loggedIn = false;
+  }
+
+  login() {
+    this.loggedIn = true;
+  }
+
+  logout() {
+    this.loggedIn = false;
+  }
+
+  isLoggedIn() {
+    return this.loggedIn;
+  }
+}
