+++ src/challenges/js/30DaysJs-30/src/exercise.js
@@ -20,22 +20,3 @@ export class MessagesProxy {
     }
   }
 }
-
-export class User {
-  constructor(name) {
-    this.name = name;
-    this.loggedIn = false;
-  }
-
-  login() {
-    this.loggedIn = true;
-  }
-
-  logout() {
-    this.loggedIn = false;
-  }
-
-  isLoggedIn() {
-    return this.loggedIn;
-  }
-}
