+++ src/challenges/js/30DaysJs-30/src/exercise.test.js
@@ -1,4 +1,5 @@
-import { MessagesProxy, User } from "./exercise";
+import { MessagesProxy } from "./exercise";
+import { User } from "./User";
 import { Messages } from "./messages";
 
 describe("tests", () => {
